@@ -1,10 +1,20 @@
 module github.com/gookit/gcli/v3
 
-go 1.16
+go 1.18
 
 require (
 	github.com/gookit/color v1.5.1
 	github.com/gookit/goutil v0.5.5
 	github.com/stretchr/testify v1.8.0
 	golang.org/x/crypto v0.0.0-20210220033148-5ea612d1eb83
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mattn/go-isatty v0.0.14 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/xo/terminfo v0.0.0-20210125001918-ca9a967f8778 // indirect
+	golang.org/x/sys v0.0.0-20210630005230-0f9fa26af87c // indirect
+	golang.org/x/term v0.0.0-20210220032956-6a3ed077a48d // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
