@@ -1,10 +1,10 @@
 package gcli
 
 import (
-	"strconv"
 	"strings"
 
 	"github.com/gookit/goutil/errorx"
+	"github.com/gookit/goutil/structs"
 	"github.com/gookit/goutil/strutil"
 )
 
@@ -12,14 +12,12 @@ import (
  * Arguments definition
  *************************************************************/
 
-// an empty argument
-var emptyArg = &Argument{}
-
 // Arguments definition
 type Arguments struct {
 	// Inherited from Command
 	name string
 	// args definition for a command.
+	//
 	// eg. {
 	// 	{"arg0", "this is first argument", false, false},
 	// 	{"arg1", "this is second argument", false, false},
@@ -28,6 +26,7 @@ type Arguments struct {
 	// record min length for args
 	// argsMinLen int
 	// record argument names and defined positional relationships
+	//
 	// {
 	// 	// name: position
 	// 	"arg0": 0,
@@ -100,11 +99,9 @@ func (ags *Arguments) ParseArgs(args []string) (err error) {
 // Usage:
 // 	cmd.AddArg("name", "description")
 // 	cmd.AddArg("name", "description", true) // required
-// 	cmd.AddArg("names", "description", true, true) // required and is array
+// 	cmd.AddArg("names", "description", true, true) // required and is arrayed
 func (ags *Arguments) AddArg(name, desc string, requiredAndArrayed ...bool) *Argument {
-	// create new argument
 	newArg := NewArgument(name, desc, requiredAndArrayed...)
-
 	return ags.AddArgument(newArg)
 }
 
@@ -116,7 +113,7 @@ func (ags *Arguments) AddArgByRule(name, rule string) *Argument {
 	newArg := NewArgument(name, mp["desc"], required)
 
 	if defVal := mp["default"]; defVal != "" {
-		newArg.Value = defVal
+		newArg.Set(defVal)
 	}
 
 	return ags.AddArgument(newArg)
@@ -200,17 +197,17 @@ func (ags *Arguments) HasArguments() bool {
 func (ags *Arguments) Arg(name string) *Argument {
 	i, ok := ags.argsIndexes[name]
 	if !ok {
-		return emptyArg
+		panicf("get not exists argument '%s'", name)
 	}
 	return ags.args[i]
 }
 
 // ArgByIndex get named arg by index
 func (ags *Arguments) ArgByIndex(i int) *Argument {
-	if i < len(ags.args) {
-		return ags.args[i]
+	if i >= len(ags.args) {
+		panicf("get not exists argument #%d", i)
 	}
-	return emptyArg
+	return ags.args[i]
 }
 
 /*************************************************************
@@ -219,31 +216,31 @@ func (ags *Arguments) ArgByIndex(i int) *Argument {
 
 // Argument a command argument definition
 type Argument struct {
+	*structs.Value
 	// Name argument name. it's required
 	Name string
 	// Desc argument description message
 	Desc string
 	// Type name. eg: string, int, array
 	// Type string
+
 	// ShowName is a name for display help. default is equals to Name.
 	ShowName string
 	// Required arg is required
 	Required bool
 	// Arrayed if is array, can allow accept multi values, and must in last.
 	Arrayed bool
-	// value *goutil.Value TODO ...
-	// value store parsed argument data. (type: string, []string)
-	Value interface{}
+
 	// Handler custom argument value handler on call GetValue()
-	Handler func(val interface{}) interface{}
-	// Validator you can add an validator, will call it on binding argument value
-	Validator func(val interface{}) (interface{}, error)
+	Handler func(val any) any
+	// Validator you can add a validator, will call it on binding argument value
+	Validator func(val any) (any, error)
 	// the argument position index in all arguments(cmd.args[index])
 	index int
 }
 
-// NewArgument quick create an new command argument
-func NewArgument(name, desc string, requiredAndArrayed ...bool) *Argument {
+// NewArg quick create a new command argument
+func NewArg(name, desc string, val any, requiredAndArrayed ...bool) *Argument {
 	var arrayed, required bool
 	if ln := len(requiredAndArrayed); ln > 0 {
 		required = requiredAndArrayed[0]
@@ -253,19 +250,50 @@ func NewArgument(name, desc string, requiredAndArrayed ...bool) *Argument {
 	}
 
 	return &Argument{
-		Name: name,
-		Desc: desc,
-		// other options
+		Name:  name,
+		Desc:  desc,
+		Value: structs.NewValue(val),
+		// other settings
 		// ShowName: name,
 		Required: required,
 		Arrayed:  arrayed,
 	}
 }
 
+// NewArgument quick create a new command argument
+func NewArgument(name, desc string, requiredAndArrayed ...bool) *Argument {
+	return NewArg(name, desc, nil, requiredAndArrayed...)
+}
+
 // SetArrayed the argument
 func (a *Argument) SetArrayed() *Argument {
 	a.Arrayed = true
 	return a
+}
+
+// WithValue to the argument
+func (a *Argument) WithValue(val any) *Argument {
+	a.Value.Set(val)
+	return a
+}
+
+// WithFn a func for config the argument
+func (a *Argument) WithFn(fn func(arg *Argument)) *Argument {
+	if fn != nil {
+		fn(a)
+	}
+	return a
+}
+
+// WithValidator set a value validator of the argument
+func (a *Argument) WithValidator(fn func(any) (any, error)) *Argument {
+	a.Validator = fn
+	return a
+}
+
+// SetValue set an validated value
+func (a *Argument) SetValue(val any) error {
+	return a.bindValue(val)
 }
 
 // Init the argument
@@ -281,7 +309,7 @@ func (a *Argument) goodArgument() string {
 	}
 
 	if !goodName.MatchString(name) {
-		panicf("the command argument name '%s' is invalid, must match: %s", name, regGoodName)
+		panicf("the argument name '%s' is invalid, must match: %s", name, regGoodName)
 	}
 
 	a.Name = name
@@ -289,34 +317,6 @@ func (a *Argument) goodArgument() string {
 		a.ShowName = name
 	}
 	return name
-}
-
-// HelpName for render help message
-func (a *Argument) HelpName() string {
-	if a.Arrayed {
-		return a.ShowName + "..."
-	}
-	return a.ShowName
-}
-
-// With a func for config the argument
-func (a *Argument) With(fn func(arg *Argument)) *Argument {
-	if fn != nil {
-		fn(a)
-	}
-
-	return a
-}
-
-// WithValidator set a value validator of the argument
-func (a *Argument) WithValidator(fn func(interface{}) (interface{}, error)) *Argument {
-	a.Validator = fn
-	return a
-}
-
-// SetValue set an validated value
-func (a *Argument) SetValue(val interface{}) error {
-	return a.bindValue(val)
 }
 
 // GetValue get value by custom handler func
@@ -329,62 +329,6 @@ func (a *Argument) GetValue() interface{} {
 	return val
 }
 
-// Int argument value to int
-func (a *Argument) Int(defVal ...int) int {
-	def := 0
-	if len(defVal) == 1 {
-		def = defVal[0]
-	}
-
-	if a.Value == nil || a.Arrayed {
-		return def
-	}
-
-	if intVal, ok := a.Value.(int); ok {
-		return intVal
-	}
-
-	if str, ok := a.Value.(string); ok {
-		val, err := strconv.Atoi(str)
-		if err == nil {
-			return val
-		}
-	}
-	return def
-}
-
-// String argument value to string
-func (a *Argument) String(defVal ...string) string {
-	def := ""
-	if len(defVal) == 1 {
-		def = defVal[0]
-	}
-
-	if a.Value == nil || a.Arrayed {
-		return def
-	}
-
-	if str, ok := a.Value.(string); ok {
-		return str
-	}
-	return def
-}
-
-// StringSplit quick split a string argument to string slice
-func (a *Argument) StringSplit(sep ...string) (ss []string) {
-	str := a.String()
-	if str == "" {
-		return
-	}
-
-	char := ","
-	if len(sep) > 0 {
-		char = sep[0]
-	}
-
-	return strings.Split(str, char)
-}
-
 // Array alias of the Strings()
 func (a *Argument) Array() (ss []string) {
 	return a.Strings()
@@ -393,9 +337,8 @@ func (a *Argument) Array() (ss []string) {
 // Strings argument value to string array, if argument isArray = true.
 func (a *Argument) Strings() (ss []string) {
 	if a.Value != nil && a.Arrayed {
-		ss = a.Value.([]string)
+		ss = a.Value.Strings()
 	}
-
 	return
 }
 
@@ -414,16 +357,27 @@ func (a *Argument) Index() int {
 	return a.index
 }
 
-// bind an value to the argument
-func (a *Argument) bindValue(val interface{}) (err error) {
-	// has validator
+// HelpName for render help message
+func (a *Argument) HelpName() string {
+	if a.Arrayed {
+		return a.ShowName + "..."
+	}
+	return a.ShowName
+}
+
+// bind a value to the argument
+func (a *Argument) bindValue(val any) (err error) {
 	if a.Validator != nil {
 		val, err = a.Validator(val)
-		if err == nil {
-			a.Value = val
+		if err != nil {
+			return
 		}
-	} else {
-		a.Value = val
 	}
+
+	if a.Handler != nil {
+		val = a.Handler(val)
+	}
+
+	a.Value.V = val
 	return
 }
